feature: Splits engine parse, compose, and price-journey into modules
