+++ packages/engine/src/fund-composition.ts
@@ -1,7 +1,10 @@
-// Domain types, read models, parse-result/warning unions, and the assurance
-// vocabulary now live in ./contracts.js. They are re-exported here verbatim so
-// the public type surface (reached via index.ts) is unchanged — exports only
-// moved, none were added or removed.
+// The engine concerns now live in their own modules: domain types/contracts in
+// ./contracts.js, the cross-concern kernel in ./internal.js, parsing in
+// ./parse.js, composition in ./compose.js, and price journeys in
+// ./price-journey.js. This file is a thin transitional barrel that keeps the
+// public surface (reached via index.ts) byte-identical while the curated
+// index.ts is a later slice — it re-exports the same types and the same three
+// public functions, none added or removed.
 export type {
   Currency,
   ExecutionMode,
@@ -46,1333 +49,5 @@ export type {
 } from "./contracts.js";
 export { validationSeverityByCode } from "./contracts.js";
 
-// Local bindings for the contracts this module references internally.
-// `NamedRecord` and `CapitalRecordBase` are package-internal — imported here but
-// intentionally not re-exported above, so they stay off the public surface.
-import type {
-  Currency,
-  ExecutionMode,
-  Direction,
-  CapitalTier,
-  Lot,
-  FundReviewData,
-  Close,
-  NamedRecord,
-  CapitalRecordBase,
-  ReserveRecord,
-  Ok,
-  InvalidJson,
-  SchemaError,
-  DuplicateReferenceId,
-  DuplicateCapitalRecordId,
-  ParseResult,
-  LoadOutcome,
-  WarningCode,
-  Warning,
-  DashboardRowKind,
-  DashboardSectionId,
-  CompositionRow,
-  DashboardFocus,
-  DashboardSummary,
-  PriceJourneyPoint,
-  PriceJourney,
-  CompositionReport,
-  DetailRecordKind,
-  DashboardDetail,
-  TierContribution,
-} from "./contracts.js";
-
-// Cross-concern helpers shared with parse/compose/price-journey live in the
-// internal kernel — imported here, never re-exported, so they stay off the
-// public surface.
-import {
-  toUsd,
-  percentOfFund,
-  roundNumber,
-  indexById,
-  requireNonEmptyString,
-  isIsoDate,
-  pushWarning,
-  schemaError,
-  isSupportedCurrency,
-  isExecutionMode,
-  isDirection,
-  isNonNegativeNumber,
-  isPositiveNumber,
-  isRecord,
-} from "./internal.js";
-
-/**
- * Reserve Lot face sums are reconciled against `amount` with a hybrid tolerance:
- * `max(absolute floor, relative fraction × amount)`. The floor absorbs cent-level
- * rounding on small balances; the relative term scales the forgiveness with the
- * balance so a hand-rounded provenance split on a large reserve does not trip the
- * warning, while real misallocation (beyond the band) still emits
- * `reserve-lot-sum-mismatch`. `amount` always stays authoritative. Both terms are
- * named and tunable here.
- */
-const RESERVE_LOT_SUM_ABS_TOLERANCE = 0.01;
-const RESERVE_LOT_SUM_REL_TOLERANCE = 0.001; // 0.1% of amount
-
-/** Hybrid reconciliation tolerance for a reserve of the given face `amount`. */
-function reserveLotSumTolerance(amount: number): number {
-  return Math.max(
-    RESERVE_LOT_SUM_ABS_TOLERANCE,
-    RESERVE_LOT_SUM_REL_TOLERANCE * Math.abs(amount),
-  );
-}
-
-/**
- * `markPrice` (the authoritative P&L input) and an instrument's latest Close
- * (display-only) are reconciled with the same hybrid shape as reserves:
- * `max(absolute floor, relative fraction × latestClose)`. The floor absorbs
- * sub-cent drift on small-priced instruments; the relative term scales the
- * forgiveness with the price so routine intraday slippage between a mark and the
- * weekly Close does not trip the warning, while a real divergence (beyond the
- * band) emits `markprice-close-mismatch`. The warning never alters valuation —
- * `markPrice` stays authoritative and Close stays display-only. Both terms are
- * named and tunable here.
- */
-const MARKPRICE_CLOSE_ABS_TOLERANCE = 0.01;
-const MARKPRICE_CLOSE_REL_TOLERANCE = 0.005; // 0.5% of the latest Close
-
-/** Hybrid coherence tolerance for a markPrice against the given latest Close. */
-function markPriceCloseTolerance(latestClose: number): number {
-  return Math.max(
-    MARKPRICE_CLOSE_ABS_TOLERANCE,
-    MARKPRICE_CLOSE_REL_TOLERANCE * Math.abs(latestClose),
-  );
-}
-
-interface CanonicalLine {
-  recordId: string;
-  recordKind: DetailRecordKind;
-  recordLabel: string;
-  portfolioId: string;
-  portfolioLabel: string;
-  tempoId: string;
-  tempoLabel: string;
-  accountId: string;
-  accountLabel: string;
-  instrumentId: string;
-  instrumentLabel: string;
-  usdValue: number;
-  costBasisUsd?: number;
-  unrealizedPnlUsd?: number;
-  tierContributions?: TierContribution[];
-}
-
-interface BuildCompositionReportOptions {
-  load?: LoadOutcome;
-}
-
-interface GroupAccumulator {
-  id: string;
-  label: string;
-  usdValue: number;
-  costBasisUsd: number;
-  unrealizedPnlUsd: number;
-}
-
-interface CanonicalState {
-  canonicalLines: CanonicalLine[];
-  warnings: Warning[];
-  excluded: CompositionReport["excluded"];
-}
-
-export function parseFundReview(input: unknown): ParseResult {
-  const parsed = parseReviewInput(input);
-  if (parsed.kind !== "ok") {
-    return parsed;
-  }
-
-  const value = parsed.value;
-  if (!isRecord(value)) {
-    return schemaError("$", "Review file must contain a JSON object.");
-  }
-
-  const requiredArrays = [
-    "portfolios",
-    "accounts",
-    "instruments",
-    "reserves",
-    "positions",
-  ] as const;
-  for (const key of requiredArrays) {
-    if (!Array.isArray(value[key])) {
-      return schemaError(key, `Review file is missing array: ${key}`);
-    }
-  }
-
-  if (!isRecord(value.fund)) {
-    return schemaError("fund", "Review file must contain fund details.");
-  }
-
-  const fundIdError = requireNonEmptyString(value.fund.id, "fund.id");
-  if (fundIdError) {
-    return fundIdError;
-  }
-
-  const fundNameError = requireNonEmptyString(value.fund.name, "fund.name");
-  if (fundNameError) {
-    return fundNameError;
-  }
-
-  if (value.fund.baseCurrency !== "USD") {
-    return {
-      kind: "unsupported-base-currency",
-      severity: "blocking",
-      baseCurrency: value.fund.baseCurrency,
-      message: "Prototype only supports a USD base Fund.",
-    };
-  }
-
-  if (!isRecord(value.review)) {
-    return schemaError("review", "Review file must contain review details.");
-  }
-
-  const asOfTypeError = requireNonEmptyString(value.review.asOf, "review.asOf");
-  if (asOfTypeError) {
-    return asOfTypeError;
-  }
-
-  if (!isIsoDate(value.review.asOf)) {
-    return {
-      kind: "invalid-as-of",
-      severity: "blocking",
-      path: "review.asOf",
-      value: value.review.asOf,
-      message: "review.asOf must be a valid YYYY-MM-DD date.",
-    };
-  }
-
-  if (!isPositiveNumber(value.review.usdMxn)) {
-    return {
-      kind: "invalid-fx-rate",
-      severity: "blocking",
-      path: "review.usdMxn",
-      value: value.review.usdMxn,
-      message: "review.usdMxn must be a positive MXN-per-USD rate.",
-    };
-  }
-
-  const portfolioError = validateNamedRecords(value.portfolios, "portfolios", "portfolio");
-  if (portfolioError) {
-    return portfolioError;
-  }
-
-  const accountError = validateAccounts(value.accounts);
-  if (accountError) {
-    return accountError;
-  }
-
-  const instrumentError = validateInstruments(value.instruments);
-  if (instrumentError) {
-    return instrumentError;
-  }
-
-  const reserveError = validateReserves(value.reserves);
-  if (reserveError) {
-    return reserveError;
-  }
-
-  const positionError = validatePositions(value.positions);
-  if (positionError) {
-    return positionError;
-  }
-
-  if (value.closes !== undefined) {
-    const closesError = validateCloses(value.closes);
-    if (closesError) {
-      return closesError;
-    }
-  }
-
-  const duplicateCapitalIdError = validateCapitalRecordIds(
-    value.reserves,
-    value.positions,
-  );
-  if (duplicateCapitalIdError) {
-    return duplicateCapitalIdError;
-  }
-
-  return {
-    kind: "ok",
-    value: value as FundReviewData,
-  };
-}
-
-export function buildCompositionReport(
-  data: FundReviewData,
-  options: BuildCompositionReportOptions = {},
-): CompositionReport {
-  const { canonicalLines, warnings, excluded } = buildCanonicalState(data);
-
-  const fundValueUsd = canonicalLines.reduce((sum, line) => sum + line.usdValue, 0);
-  if (fundValueUsd <= 0) {
-    pushWarning(
-      warnings,
-      "non-positive-fund-value",
-      "Canonical live Fund value is not positive; percent-of-Fund values may be misleading.",
-    );
-  }
-
-  const portfolioRows = groupLines(
-    canonicalLines,
-    "portfolio",
-    (line) => line.portfolioId,
-    (line) => line.portfolioLabel,
-    fundValueUsd,
-  );
-  const tempoRows = groupLines(
-    canonicalLines,
-    "tempo",
-    (line) => line.tempoId,
-    (line) => line.tempoLabel,
-    fundValueUsd,
-  );
-  const accountRows = groupLines(
-    canonicalLines,
-    "account",
-    (line) => line.accountId,
-    (line) => line.accountLabel,
-    fundValueUsd,
-  );
-  const instrumentRows = groupLines(
-    canonicalLines,
-    "instrument",
-    (line) => line.instrumentId,
-    (line) => line.instrumentLabel,
-    fundValueUsd,
-  );
-  const tierRows = groupTierLines(canonicalLines, fundValueUsd);
-  const totalUnrealizedPnlUsd = canonicalLines.reduce(
-    (sum, line) => sum + (line.unrealizedPnlUsd ?? 0),
-    0,
-  );
-  const priceJourneys = buildPriceJourneys(data, warnings);
-
-  return {
-    totals: {
-      baseCurrency: "USD",
-      fundValueUsd,
-      usdMxn: data.review.usdMxn,
-    },
-    dashboard: {
-      summary: {
-        fundName: data.fund.name,
-        asOf: data.review.asOf,
-        fundValueUsd,
-        usdMxn: data.review.usdMxn,
-        totalUnrealizedPnlUsd,
-        ...optionalSummaryFocus("largestPortfolio", toFocus(portfolioRows[0])),
-        ...optionalSummaryFocus("largestTempo", toFocus(tempoRows[0])),
-        ...optionalSummaryFocus("largestAccount", toFocus(accountRows[0])),
-        ...optionalSummaryFocus("largestInstrument", toFocus(instrumentRows[0])),
-        ...optionalSummaryFocus(
-          "reserve",
-          toFocus(tempoRows.find((row) => row.id === "tempo:Reserve")),
-        ),
-        dataSafety: {
-          nonLiveExcluded: excluded.nonLive,
-          invalidExcluded: excluded.invalid,
-          shortDeferredExcluded: excluded.shortDeferred,
-          hasWarnings: warnings.length > 0,
-        },
-      },
-      sections: [
-        {
-          id: "portfolios",
-          title: "Portfolio Composition",
-          rows: portfolioRows,
-        },
-        {
-          id: "tempos",
-          title: "Tempo Composition",
-          rows: tempoRows,
-        },
-        {
-          id: "accounts",
-          title: "Account Composition",
-          rows: accountRows,
-        },
-        {
-          id: "instruments",
-          title: "Instrument Composition",
-          rows: instrumentRows,
-        },
-        {
-          id: "tiers",
-          title: "Capital Tier Composition",
-          rows: tierRows,
-        },
-      ],
-    },
-    priceJourneys,
-    warnings,
-    excluded,
-    load: options.load ?? { status: "loaded" },
-  };
-}
-
-function buildPriceJourneys(
-  data: FundReviewData,
-  warnings: Warning[],
-): PriceJourney[] {
-  if (!Array.isArray(data.closes) || data.closes.length === 0) {
-    return [];
-  }
-
-  const instruments = indexById(data.instruments, "instrument");
-  const byInstrument = new Map<string, PriceJourneyPoint[]>();
-  for (const close of data.closes) {
-    // A Close referencing an unknown Instrument or carrying an invalid scalar is
-    // dropped from its journey, but the drop is surfaced via `skipped-close`
-    // rather than silently swallowed — and a bad anchor never blocks its valid
-    // siblings, which still render the journey for that instrument.
-    const instrument = instruments.get(close.instrumentId);
-    if (!instrument) {
-      pushWarning(
-        warnings,
-        "skipped-close",
-        `Close for unknown Instrument ${close.instrumentId} (as of ${String(close.asOf)}) was skipped; valid anchors still render.`,
-        close.instrumentId,
-      );
-      continue;
-    }
-    if (!isIsoDate(close.asOf) || !isNonNegativeNumber(close.price)) {
-      pushWarning(
-        warnings,
-        "skipped-close",
-        `Close for Instrument ${close.instrumentId} (as of ${String(close.asOf)}) has an invalid date or price and was skipped; valid anchors still render.`,
-        close.instrumentId,
-      );
-      continue;
-    }
-    const points = byInstrument.get(close.instrumentId) ?? [];
-    points.push({ asOf: close.asOf, price: close.price });
-    byInstrument.set(close.instrumentId, points);
-  }
-
-  const journeys: PriceJourney[] = [];
-  for (const [instrumentId, points] of byInstrument) {
-    if (points.length < 2) continue; // a single anchor is not a journey
-    const sorted = [...points].sort((a, b) => a.asOf.localeCompare(b.asOf));
-    const firstPrice = sorted[0]!.price;
-    const latestPrice = sorted[sorted.length - 1]!.price;
-    const instrument = instruments.get(instrumentId)!;
-    journeys.push({
-      instrumentId,
-      label: `${instrument.symbol} (${instrument.name})`,
-      currency: instrument.currency,
-      points: sorted,
-      firstPrice,
-      latestPrice,
-      changeAbs: latestPrice - firstPrice,
-      changePct:
-        firstPrice === 0
-          ? 0
-          : roundNumber(((latestPrice - firstPrice) / firstPrice) * 100, 12),
-    });
-  }
-
-  return journeys.sort(
-    (a, b) => b.points.length - a.points.length || a.label.localeCompare(b.label),
-  );
-}
-
-/**
- * Latest valid Close per instrument, keyed by instrumentId. Only Closes with a
- * valid date and scalar are considered (an invalid scalar is surfaced as
- * `skipped-close` in {@link buildPriceJourneys}, not used as a coherence anchor).
- * "Latest" is the lexicographically greatest ISO `asOf`. Display-only: the result
- * feeds the non-blocking `markprice-close-mismatch` check and never valuation.
- */
-function latestCloseByInstrument(
-  closes: Close[] | undefined,
-): Map<string, Close> {
-  const latest = new Map<string, Close>();
-  if (!Array.isArray(closes)) return latest;
-  for (const close of closes) {
-    if (!isIsoDate(close.asOf) || !isNonNegativeNumber(close.price)) continue;
-    const existing = latest.get(close.instrumentId);
-    if (!existing || close.asOf.localeCompare(existing.asOf) > 0) {
-      latest.set(close.instrumentId, close);
-    }
-  }
-  return latest;
-}
-
-export function buildDashboardDetail(
-  data: FundReviewData,
-  report: CompositionReport,
-  rowId: string,
-): DashboardDetail | undefined {
-  const row = findDashboardRow(report, rowId);
-  if (!row || row.kind === "instrument" || row.kind === "tier") {
-    return undefined;
-  }
-
-  const { canonicalLines } = buildCanonicalState(data);
-  return {
-    rowId: row.id,
-    kind: row.kind,
-    label: row.label,
-    rows: detailLinesForRow(canonicalLines, row).map((line) => ({
-      recordId: line.recordId,
-      kind: line.recordKind,
-      recordLabel: line.recordLabel,
-      portfolioLabel: line.portfolioLabel,
-      tempoLabel: line.tempoLabel,
-      accountLabel: line.accountLabel,
-      usdValue: line.usdValue,
-      ...(line.tierContributions
-        ? { tierContributions: line.tierContributions }
-        : {}),
-    })),
-  };
-}
-
-function buildCanonicalState(data: FundReviewData): CanonicalState {
-  const warnings: Warning[] = [];
-  const portfolios = indexById(data.portfolios, "portfolio");
-  const accounts = indexById(data.accounts, "account");
-  const instruments = indexById(data.instruments, "instrument");
-  const latestCloses = latestCloseByInstrument(data.closes);
-  const canonicalLines: CanonicalLine[] = [];
-  const excluded = {
-    nonLive: 0,
-    invalid: 0,
-    shortDeferred: 0,
-  };
-
-  for (const reserve of data.reserves) {
-    if (!isExecutionMode(reserve.executionMode)) {
-      pushWarning(
-        warnings,
-        "unsupported-execution-mode",
-        `Reserve ${reserve.id} uses unsupported Execution Mode ${String(reserve.executionMode)} and was excluded.`,
-        reserve.id,
-      );
-      excluded.invalid += 1;
-      continue;
-    }
-
-    if (reserve.executionMode !== "live") {
-      excluded.nonLive += 1;
-      continue;
-    }
-
-    if (!isSupportedCurrency(reserve.currency)) {
-      pushWarning(
-        warnings,
-        "unsupported-currency",
-        `Reserve ${reserve.id} uses unsupported Currency ${String(reserve.currency)} and was excluded.`,
-        reserve.id,
-      );
-      excluded.invalid += 1;
-      continue;
-    }
-
-    const reserveHasMissingReference = validateCapitalBase(
-      reserve,
-      portfolios,
-      accounts,
-      warnings,
-    );
-    if (reserveHasMissingReference) {
-      excluded.invalid += 1;
-      continue;
-    }
-
-    const reserveAccount = accounts.get(reserve.accountId);
-    if (reserveAccount && reserveAccount.currency !== reserve.currency) {
-      pushWarning(
-        warnings,
-        "currency-mismatch",
-        `Reserve ${reserve.id} currency ${reserve.currency} does not match Account ${reserve.accountId} currency ${reserveAccount.currency} and was excluded.`,
-        reserve.id,
-      );
-      excluded.invalid += 1;
-      continue;
-    }
-
-    if (!isNonNegativeNumber(reserve.amount)) {
-      pushWarning(
-        warnings,
-        "invalid-amount",
-        `Reserve ${reserve.id} has invalid amount and was excluded.`,
-        reserve.id,
-      );
-      excluded.invalid += 1;
-      continue;
-    }
-
-    // Cash Lots attribute the reserve's face to Capital Tiers (cost == value,
-    // Price P&L == 0). `amount` stays the line's authoritative USD value; the
-    // tier contributions are an independent overlay consumed only by the tier
-    // rollup, so untiered reserves and every other section stay unchanged.
-    const tierContributions = buildReserveTierContributions(
-      reserve,
-      data.review.usdMxn,
-      warnings,
-    );
-
-    canonicalLines.push({
-      recordId: reserve.id,
-      recordKind: "reserve",
-      recordLabel: "Reserve",
-      portfolioId: reserve.portfolioId,
-      portfolioLabel: portfolios.get(reserve.portfolioId)?.name ?? reserve.portfolioId,
-      tempoId: reserve.tempo,
-      tempoLabel: reserve.tempo,
-      accountId: reserve.accountId,
-      accountLabel: accountLabel(reserveAccount, reserve.accountId),
-      instrumentId: "reserve",
-      instrumentLabel: "Reserve",
-      usdValue: toUsd(reserve.amount, reserve.currency, data.review.usdMxn),
-      ...(tierContributions ? { tierContributions } : {}),
-    });
-  }
-
-  for (const position of data.positions) {
-    if (!isExecutionMode(position.executionMode)) {
-      pushWarning(
-        warnings,
-        "unsupported-execution-mode",
-        `Position ${position.id} uses unsupported Execution Mode ${String(position.executionMode)} and was excluded.`,
-        position.id,
-      );
-      excluded.invalid += 1;
-      continue;
-    }
-
-    if (position.executionMode !== "live") {
-      excluded.nonLive += 1;
-      continue;
-    }
-
-    if (!isSupportedCurrency(position.currency)) {
-      pushWarning(
-        warnings,
-        "unsupported-currency",
-        `Position ${position.id} uses unsupported Currency ${String(position.currency)} and was excluded.`,
-        position.id,
-      );
-      excluded.invalid += 1;
-      continue;
-    }
-
-    if (!isDirection(position.direction)) {
-      pushWarning(
-        warnings,
-        "unsupported-direction",
-        `Position ${position.id} uses unsupported Direction ${String(position.direction)} and was excluded.`,
-        position.id,
-      );
-      excluded.invalid += 1;
-      continue;
-    }
-
-    if (position.direction === "short") {
-      excluded.shortDeferred += 1;
-      continue;
-    }
-
-    let invalidRecord = validateCapitalBase(position, portfolios, accounts, warnings);
-
-    const instrument = instruments.get(position.instrumentId);
-    if (!instrument) {
-      pushWarning(
-        warnings,
-        "missing-instrument",
-        `Position ${position.id} references missing Instrument ${position.instrumentId} and was excluded.`,
-        position.id,
-      );
-      invalidRecord = true;
-    }
-
-    const account = accounts.get(position.accountId);
-    if (account && account.currency !== position.currency) {
-      pushWarning(
-        warnings,
-        "currency-mismatch",
-        `Position ${position.id} currency ${position.currency} does not match Account ${position.accountId} currency ${account.currency} and was excluded.`,
-        position.id,
-      );
-      invalidRecord = true;
-    }
-
-    if (instrument && instrument.currency !== position.currency) {
-      pushWarning(
-        warnings,
-        "currency-mismatch",
-        `Position ${position.id} currency ${position.currency} does not match Instrument ${position.instrumentId} currency ${instrument.currency} and was excluded.`,
-        position.id,
-      );
-      invalidRecord = true;
-    }
-
-    if (invalidRecord) {
-      excluded.invalid += 1;
-      continue;
-    }
-
-    const lots = position.lots;
-    const invalidNumericFields: string[] = [];
-    if (!isNonNegativeNumber(position.markPrice)) {
-      invalidNumericFields.push("markPrice");
-    }
-    if (lots.length === 0) {
-      invalidNumericFields.push("lots");
-    }
-    for (const lot of lots) {
-      if (!isNonNegativeNumber(lot.quantity)) invalidNumericFields.push("quantity");
-      if (!isNonNegativeNumber(lot.cost)) invalidNumericFields.push("cost");
-      if (lot.entryFx !== undefined && !isPositiveNumber(lot.entryFx)) {
-        invalidNumericFields.push("entryFx");
-      }
-    }
-    if (invalidNumericFields.length > 0) {
-      pushWarning(
-        warnings,
-        "invalid-position-number",
-        `Position ${position.id} has invalid ${[...new Set(invalidNumericFields)].join(", ")} and was excluded.`,
-        position.id,
-      );
-      excluded.invalid += 1;
-      continue;
-    }
-
-    // Market value converts at the current review FX; each Lot's cost basis
-    // converts at its own entry FX (falling back to the review FX). P&L is the
-    // per-Lot join of (quantity, cost, tier, entryFx) against one instrument
-    // markPrice, then aggregated by Capital Tier.
-    const reviewFx = data.review.usdMxn;
-    let marketValueUsd = 0;
-    let costBasisUsd = 0;
-    const tierTotals = new Map<CapitalTier, TierContribution>();
-    for (const lot of lots) {
-      const lotMarketUsd = toUsd(lot.quantity * position.markPrice, position.currency, reviewFx);
-      const lotCostUsd = toUsd(lot.quantity * lot.cost, position.currency, lot.entryFx ?? reviewFx);
-      marketValueUsd += lotMarketUsd;
-      costBasisUsd += lotCostUsd;
-      const existing = tierTotals.get(lot.tier) ?? {
-        tier: lot.tier,
-        usdValue: 0,
-        costBasisUsd: 0,
-        unrealizedPnlUsd: 0,
-      };
-      existing.usdValue += lotMarketUsd;
-      existing.costBasisUsd += lotCostUsd;
-      existing.unrealizedPnlUsd += lotMarketUsd - lotCostUsd;
-      tierTotals.set(lot.tier, existing);
-    }
-
-    canonicalLines.push({
-      recordId: position.id,
-      recordKind: "position",
-      recordLabel: `${instrument!.symbol} (${instrument!.name})`,
-      portfolioId: position.portfolioId,
-      portfolioLabel: portfolios.get(position.portfolioId)?.name ?? position.portfolioId,
-      tempoId: position.tempo,
-      tempoLabel: position.tempo,
-      accountId: position.accountId,
-      accountLabel: accountLabel(account, position.accountId),
-      instrumentId: position.instrumentId,
-      instrumentLabel: `${instrument!.symbol} (${instrument!.name})`,
-      usdValue: marketValueUsd,
-      costBasisUsd,
-      unrealizedPnlUsd: marketValueUsd - costBasisUsd,
-      tierContributions: [...tierTotals.values()],
-    });
-
-    // Per-position coherence check (non-blocking, valuation already committed
-    // above): a position's authoritative `markPrice` and its instrument's latest
-    // display-only Close should agree within tolerance. Running per-position
-    // incidentally surfaces cross-position divergence of the same instrument.
-    // markPrice and Close share the instrument's native units (currency is
-    // reconciled above), so they compare directly without FX conversion.
-    const latestClose = latestCloses.get(position.instrumentId);
-    if (
-      latestClose &&
-      Math.abs(position.markPrice - latestClose.price) >
-        markPriceCloseTolerance(latestClose.price)
-    ) {
-      pushWarning(
-        warnings,
-        "markprice-close-mismatch",
-        `Position ${position.id} markPrice ${position.markPrice} disagrees with Instrument ${position.instrumentId} latest Close ${latestClose.price} (as of ${latestClose.asOf}) beyond tolerance; markPrice stays the authoritative P&L input and Close stays display-only.`,
-        position.id,
-      );
-    }
-  }
-
-  return {
-    canonicalLines,
-    warnings,
-    excluded,
-  };
-}
-
-function findDashboardRow(
-  report: CompositionReport,
-  rowId: string,
-): CompositionRow | undefined {
-  for (const section of report.dashboard.sections) {
-    const row = section.rows.find((candidate) => candidate.id === rowId);
-    if (row) {
-      return row;
-    }
-  }
-  return undefined;
-}
-
-function detailLinesForRow(
-  canonicalLines: CanonicalLine[],
-  row: CompositionRow,
-): CanonicalLine[] {
-  const rawId = row.id.slice(row.id.indexOf(":") + 1);
-
-  if (row.kind === "portfolio") {
-    return canonicalLines.filter(
-      (line) => line.recordKind === "position" && line.portfolioId === rawId,
-    );
-  }
-
-  if (row.kind === "tempo") {
-    return canonicalLines.filter((line) => line.tempoId === rawId);
-  }
-
-  const accountLines = canonicalLines.filter((line) => line.accountId === rawId);
-  if (accountLines.some((line) => line.recordKind === "reserve")) {
-    return accountLines;
-  }
-
-  return accountLines.filter((line) => line.recordKind === "position");
-}
-
-function toFocus(row: CompositionRow | undefined): DashboardFocus | undefined {
-  if (!row) return undefined;
-  return {
-    rowId: row.id,
-    kind: row.kind,
-    label: row.label,
-    usdValue: row.usdValue,
-    percentOfFund: row.percentOfFund,
-  };
-}
-
-function optionalSummaryFocus<
-  Key extends
-    | "largestPortfolio"
-    | "largestTempo"
-    | "largestAccount"
-    | "largestInstrument"
-    | "reserve",
->(
-  key: Key,
-  value: DashboardSummary[Key],
-): Pick<DashboardSummary, Key> | Record<string, never> {
-  return value === undefined ? {} : ({ [key]: value } as Pick<DashboardSummary, Key>);
-}
-
-function groupLines(
-  lines: CanonicalLine[],
-  kind: DashboardRowKind,
-  idSelector: (line: CanonicalLine) => string,
-  labelSelector: (line: CanonicalLine) => string,
-  fundValueUsd: number,
-): CompositionRow[] {
-  const rows = new Map<string, GroupAccumulator>();
-
-  for (const line of lines) {
-    const rawId = idSelector(line);
-    const label = labelSelector(line);
-    const id = `${kind}:${rawId}`;
-    const existing = rows.get(id) ?? {
-      id,
-      label,
-      usdValue: 0,
-      costBasisUsd: 0,
-      unrealizedPnlUsd: 0,
-    };
-    existing.usdValue += line.usdValue;
-    existing.costBasisUsd += line.costBasisUsd ?? 0;
-    existing.unrealizedPnlUsd += line.unrealizedPnlUsd ?? 0;
-    rows.set(id, existing);
-  }
-
-  return [...rows.values()]
-    .map((row) => {
-      const result: CompositionRow = {
-        id: row.id,
-        kind,
-        label: row.label,
-        usdValue: row.usdValue,
-        percentOfFund: percentOfFund(row.usdValue, fundValueUsd),
-      };
-
-      if (row.costBasisUsd !== 0) {
-        result.costBasisUsd = row.costBasisUsd;
-      }
-      if (row.unrealizedPnlUsd !== 0) {
-        result.unrealizedPnlUsd = row.unrealizedPnlUsd;
-      }
-
-      return result;
-    })
-    .sort(
-      (a, b) =>
-        Math.abs(b.usdValue) - Math.abs(a.usdValue) || a.label.localeCompare(b.label),
-    );
-}
-
-function buildReserveTierContributions(
-  reserve: ReserveRecord,
-  reviewFx: number,
-  warnings: Warning[],
-): TierContribution[] | undefined {
-  const lots = reserve.lots;
-  if (!Array.isArray(lots) || lots.length === 0) {
-    return undefined; // untiered = excluded from the tier rollup (back-compat)
-  }
-
-  // A bad scalar slipped past schema typing leaves the reserve untiered rather
-  // than poisoning the rollup with NaN; `amount` still counts toward the fund.
-  // The drop is surfaced, never silent.
-  if (lots.some((lot) => !isNonNegativeNumber(lot.quantity))) {
-    pushWarning(
-      warnings,
-      "invalid-reserve-lot-quantity",
-      `Reserve ${reserve.id} has a Lot with a negative or non-finite quantity; the Reserve stays untiered and amount stays authoritative.`,
-      reserve.id,
-    );
-    return undefined;
-  }
-
-  const tierTotals = new Map<CapitalTier, TierContribution>();
-  let faceSum = 0;
-  for (const lot of lots) {
-    const lotUsd = toUsd(lot.quantity, reserve.currency, reviewFx);
-    faceSum += lot.quantity;
-    const existing = tierTotals.get(lot.tier) ?? {
-      tier: lot.tier,
-      usdValue: 0,
-      costBasisUsd: 0,
-      unrealizedPnlUsd: 0,
-    };
-    existing.usdValue += lotUsd;
-    existing.costBasisUsd += lotUsd; // cash: cost == face, so Price P&L == 0
-    tierTotals.set(lot.tier, existing);
-  }
-
-  if (Math.abs(faceSum - reserve.amount) > reserveLotSumTolerance(reserve.amount)) {
-    pushWarning(
-      warnings,
-      "reserve-lot-sum-mismatch",
-      `Reserve ${reserve.id} Lot tiers sum to ${faceSum} ${reserve.currency} but amount is ${reserve.amount} ${reserve.currency}; amount stays authoritative and the tier split is taken as-given.`,
-      reserve.id,
-    );
-  }
-
-  return [...tierTotals.values()];
-}
-
-function groupTierLines(
-  lines: CanonicalLine[],
-  fundValueUsd: number,
-): CompositionRow[] {
-  const rows = new Map<string, GroupAccumulator>();
-
-  for (const line of lines) {
-    if (!line.tierContributions) continue;
-    for (const contribution of line.tierContributions) {
-      const id = `tier:${contribution.tier}`;
-      const existing = rows.get(id) ?? {
-        id,
-        label: contribution.tier,
-        usdValue: 0,
-        costBasisUsd: 0,
-        unrealizedPnlUsd: 0,
-      };
-      existing.usdValue += contribution.usdValue;
-      existing.costBasisUsd += contribution.costBasisUsd;
-      existing.unrealizedPnlUsd += contribution.unrealizedPnlUsd;
-      rows.set(id, existing);
-    }
-  }
-
-  return [...rows.values()]
-    .map((row) => {
-      const result: CompositionRow = {
-        id: row.id,
-        kind: "tier",
-        label: row.label,
-        usdValue: row.usdValue,
-        percentOfFund: percentOfFund(row.usdValue, fundValueUsd),
-      };
-      if (row.costBasisUsd !== 0) {
-        result.costBasisUsd = row.costBasisUsd;
-      }
-      if (row.unrealizedPnlUsd !== 0) {
-        result.unrealizedPnlUsd = row.unrealizedPnlUsd;
-      }
-      return result;
-    })
-    .sort(
-      (a, b) =>
-        Math.abs(b.usdValue) - Math.abs(a.usdValue) || a.label.localeCompare(b.label),
-    );
-}
-
-function accountLabel(
-  account: (NamedRecord & { platform: string }) | undefined,
-  fallback: string,
-): string {
-  return account ? `${account.platform}: ${account.name}` : fallback;
-}
-
-function validateCapitalBase(
-  record: CapitalRecordBase,
-  portfolios: Map<string, NamedRecord>,
-  accounts: Map<string, NamedRecord>,
-  warnings: Warning[],
-): boolean {
-  let hasMissingReference = false;
-  if (!portfolios.has(record.portfolioId)) {
-    pushWarning(
-      warnings,
-      "missing-portfolio",
-      `${record.id} references missing Portfolio ${record.portfolioId} and was excluded.`,
-      record.id,
-    );
-    hasMissingReference = true;
-  }
-  if (!accounts.has(record.accountId)) {
-    pushWarning(
-      warnings,
-      "missing-account",
-      `${record.id} references missing Account ${record.accountId} and was excluded.`,
-      record.id,
-    );
-    hasMissingReference = true;
-  }
-  return hasMissingReference;
-}
-
-function validateNamedRecords(
-  value: unknown,
-  path: string,
-  recordType: DuplicateReferenceId["recordType"],
-): ParseResult | undefined {
-  if (!Array.isArray(value)) {
-    return schemaError(path, `Review file is missing array: ${path}`);
-  }
-
-  const ids = new Set<string>();
-  for (const [index, record] of value.entries()) {
-    const itemPath = `${path}[${index}]`;
-    if (!isRecord(record)) {
-      return schemaError(itemPath, `${itemPath} must be an object.`);
-    }
-
-    const idError = requireNonEmptyString(record.id, `${itemPath}.id`);
-    if (idError) {
-      return idError;
-    }
-
-    const nameError = requireNonEmptyString(record.name, `${itemPath}.name`);
-    if (nameError) {
-      return nameError;
-    }
-
-    const id = record.id as string;
-    if (ids.has(id)) {
-      return {
-        kind: "duplicate-reference-id",
-        severity: "blocking",
-        recordType,
-        id,
-        message: `Duplicate ${recordType} id: ${id}`,
-      };
-    }
-
-    ids.add(id);
-  }
-
-  return undefined;
-}
-
-function validateAccounts(value: unknown): ParseResult | undefined {
-  const namedError = validateNamedRecords(value, "accounts", "account");
-  if (namedError) {
-    return namedError;
-  }
-
-  const accounts = value as Array<Record<string, unknown>>;
-  for (const [index, account] of accounts.entries()) {
-    const itemPath = `accounts[${index}]`;
-    const platformError = requireNonEmptyString(account.platform, `${itemPath}.platform`);
-    if (platformError) {
-      return platformError;
-    }
-
-    const currencyError = requireNonEmptyString(account.currency, `${itemPath}.currency`);
-    if (currencyError) {
-      return currencyError;
-    }
-  }
-
-  return undefined;
-}
-
-function validateInstruments(value: unknown): ParseResult | undefined {
-  const namedError = validateNamedRecords(value, "instruments", "instrument");
-  if (namedError) {
-    return namedError;
-  }
-
-  const instruments = value as Array<Record<string, unknown>>;
-  for (const [index, instrument] of instruments.entries()) {
-    const itemPath = `instruments[${index}]`;
-    const symbolError = requireNonEmptyString(instrument.symbol, `${itemPath}.symbol`);
-    if (symbolError) {
-      return symbolError;
-    }
-
-    const currencyError = requireNonEmptyString(instrument.currency, `${itemPath}.currency`);
-    if (currencyError) {
-      return currencyError;
-    }
-  }
-
-  return undefined;
-}
-
-function validateReserves(value: unknown): ParseResult | undefined {
-  if (!Array.isArray(value)) {
-    return schemaError("reserves", "Review file is missing array: reserves");
-  }
-
-  const reserves = value as Array<Record<string, unknown>>;
-  for (const [index, reserve] of reserves.entries()) {
-    const itemPath = `reserves[${index}]`;
-    const baseError = validateCapitalRecordShape(reserve, itemPath);
-    if (baseError) {
-      return baseError;
-    }
-
-    if (typeof reserve.amount !== "number") {
-      return schemaError(`${itemPath}.amount`, `${itemPath}.amount must be a number.`);
-    }
-
-    if (reserve.lots !== undefined) {
-      const lotsError = validateLots(reserve.lots, itemPath, { requireCost: false });
-      if (lotsError) {
-        return lotsError;
-      }
-    }
-  }
-
-  return undefined;
-}
-
-function validatePositions(value: unknown): ParseResult | undefined {
-  if (!Array.isArray(value)) {
-    return schemaError("positions", "Review file is missing array: positions");
-  }
-
-  const positions = value as Array<Record<string, unknown>>;
-  for (const [index, position] of positions.entries()) {
-    const itemPath = `positions[${index}]`;
-    const baseError = validateCapitalRecordShape(position, itemPath);
-    if (baseError) {
-      return baseError;
-    }
-
-    const instrumentIdError = requireNonEmptyString(
-      position.instrumentId,
-      `${itemPath}.instrumentId`,
-    );
-    if (instrumentIdError) {
-      return instrumentIdError;
-    }
-
-    const directionError = requireNonEmptyString(position.direction, `${itemPath}.direction`);
-    if (directionError) {
-      return directionError;
-    }
-
-    if (typeof position.markPrice !== "number") {
-      return schemaError(
-        `${itemPath}.markPrice`,
-        `${itemPath}.markPrice must be a number.`,
-      );
-    }
-
-    // `lots` is the only cost-carrier; a Position must carry at least one Lot.
-    const lotsError = validateLots(position.lots, itemPath);
-    if (lotsError) {
-      return lotsError;
-    }
-  }
-
-  return undefined;
-}
-
-function validateCloses(value: unknown): ParseResult | undefined {
-  if (!Array.isArray(value)) {
-    return schemaError("closes", "closes must be an array when present.");
-  }
-
-  for (const [index, close] of value.entries()) {
-    const itemPath = `closes[${index}]`;
-    if (!isRecord(close)) {
-      return schemaError(itemPath, `${itemPath} must be an object.`);
-    }
-
-    const instrumentIdError = requireNonEmptyString(
-      close.instrumentId,
-      `${itemPath}.instrumentId`,
-    );
-    if (instrumentIdError) {
-      return instrumentIdError;
-    }
-
-    const asOfError = requireNonEmptyString(close.asOf, `${itemPath}.asOf`);
-    if (asOfError) {
-      return asOfError;
-    }
-
-    if (typeof close.price !== "number") {
-      return schemaError(`${itemPath}.price`, `${itemPath}.price must be a number.`);
-    }
-  }
-
-  return undefined;
-}
-
-function validateLots(
-  value: unknown,
-  path: string,
-  options: { requireCost?: boolean } = {},
-): ParseResult | undefined {
-  // Position Lots carry a cost; cash (Reserve) Lots are degenerate — value ==
-  // face — so they omit `cost`/`entryFx` and only carry `quantity` + `tier`.
-  const requireCost = options.requireCost ?? true;
-  if (!Array.isArray(value) || value.length === 0) {
-    return schemaError(`${path}.lots`, `${path}.lots must be a non-empty array.`);
-  }
-
-  for (const [index, lot] of value.entries()) {
-    const lotPath = `${path}.lots[${index}]`;
-    if (!isRecord(lot)) {
-      return schemaError(lotPath, `${lotPath} must be an object.`);
-    }
-
-    const numericFields = requireCost
-      ? (["quantity", "cost"] as const)
-      : (["quantity"] as const);
-    for (const field of numericFields) {
-      if (typeof lot[field] !== "number") {
-        return schemaError(
-          `${lotPath}.${field}`,
-          `${lotPath}.${field} must be a number.`,
-        );
-      }
-    }
-
-    if (lot.tier !== "c1" && lot.tier !== "c2" && lot.tier !== "c3") {
-      return schemaError(
-        `${lotPath}.tier`,
-        `${lotPath}.tier must be one of c1, c2, c3.`,
-      );
-    }
-
-    if (lot.entryFx !== undefined && typeof lot.entryFx !== "number") {
-      return schemaError(
-        `${lotPath}.entryFx`,
-        `${lotPath}.entryFx must be a number when present.`,
-      );
-    }
-  }
-
-  return undefined;
-}
-
-function validateCapitalRecordShape(
-  value: unknown,
-  path: string,
-): ParseResult | undefined {
-  if (!isRecord(value)) {
-    return schemaError(path, `${path} must be an object.`);
-  }
-
-  for (const field of [
-    "id",
-    "portfolioId",
-    "tempo",
-    "executionMode",
-    "accountId",
-    "currency",
-  ] as const) {
-    const error = requireNonEmptyString(value[field], `${path}.${field}`);
-    if (error) {
-      return error;
-    }
-  }
-
-  return undefined;
-}
-
-function validateCapitalRecordIds(
-  reserves: unknown[],
-  positions: unknown[],
-): DuplicateCapitalRecordId | undefined {
-  const ids = new Set<string>();
-
-  for (const record of [...reserves, ...positions]) {
-    if (!isRecord(record) || typeof record.id !== "string") {
-      continue;
-    }
-
-    if (ids.has(record.id)) {
-      return {
-        kind: "duplicate-capital-record-id",
-        severity: "blocking",
-        id: record.id,
-        message: `Duplicate capital record id: ${record.id}`,
-      };
-    }
-
-    ids.add(record.id);
-  }
-
-  return undefined;
-}
-
-function parseReviewInput(input: unknown): Ok | InvalidJson {
-  if (typeof input !== "string") {
-    return {
-      kind: "ok",
-      value: input as FundReviewData,
-    };
-  }
-
-  try {
-    return {
-      kind: "ok",
-      value: JSON.parse(input) as FundReviewData,
-    };
-  } catch (error) {
-    const detail = error instanceof Error ? error.message : String(error);
-    return {
-      kind: "invalid-json",
-      severity: "blocking",
-      message: "Review file contains invalid JSON.",
-      detail,
-    };
-  }
-}
-
+export { parseFundReview } from "./parse.js";
+export { buildCompositionReport, buildDashboardDetail } from "./compose.js";
